Refactor App.tsx into smaller components
This refactoring breaks down the main App.tsx file into several smaller, more manageable components.

- A new `MainPage.tsx` component now holds the main application logic and state.
- The header has been extracted into a `Header.tsx` component.
- The question display area is now a `QuestionDisplay.tsx` component.
- The action buttons (Shuffle, Next, etc.) are in an `ActionButtons.tsx` component.
- `App.tsx` is now a simple entry point that renders the `MainPage` component.

This change improves code organization and maintainability without altering functionality.
